// code/backend/api/src/core/config/constants.ts
export enum UserRole {
    USER = "USER",
    ADMIN = "ADMIN"
  }
  
  export enum QuestionType {
    MULTIPLE_ANSWERS = "MULTIPLE_ANSWERS",
    TRUE_FALSE = "TRUE_FALSE",
    NORMAL = "NORMAL"
  }
  
  export enum QuestionSection {
    TECHNICAL = "TECHNICAL",
    CRITICAL = "CRITICAL",
    OTHER = "OTHER"
  }
  
  export enum Answer {
    A = "A",
    B = "B",
    C = "C",
    D = "D"
  }
  
  export enum UserAnswerStatus {
    CORRECT = "CORRECT",
    WRONG = "WRONG",
    EMPTY = "EMPTY"
  }
  